+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Version numbers are based on [Semantic Versioning](https://semver.org).
 
 - Log messages from [LogService](src/classes/Service/LogService.php) in console, based on the current verbosity level
 - Scheduler service to be used for scheduling tasks for execution in the future using [SchedulerService](src/classes/Service/SchedulerService.php) class
+- SlackController: Restore previous user status after the lunch break has finished
 
 ### Fixed